requirements.txt
# Core
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
pydantic[email]
email-validator

# LangChain / LangGraph
langgraph
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-qdrant
langchain-text-splitters

# LLM + Embeddings
groq
sentence-transformers

# Vector DB
qdrant-client

# PDF
pypdf

# Database
sqlalchemy
asyncpg
psycopg2-binary

# Auth
passlib==1.7.4
bcrypt==4.0.1
pyjwt

# Supabase
supabase

# Utilities
requests
tenacity